1. add O3 flags when compiling the utility library; 2. replace all float variables with double variables; 3. fix a bug in util.cpp

+++ src/test_util.cpp
@@ -6,8 +6,8 @@
 using std::vector;
 
 void test_effective_sinr(){
-  vector<float> rbgs_sinrs{7.53, 7.53, 7.53};
-  float effective_sinr = get_effective_sinr(rbgs_sinrs);
+  vector<double> rbgs_sinrs{7.53, 7.53, 7.53};
+  double effective_sinr = get_effective_sinr(rbgs_sinrs);
   std::cerr << "effective_sinr: " << effective_sinr << std::endl;
   int mcs = get_mcs_from_cqi(get_cqi_from_sinr(effective_sinr));
   // cqi = 6
